after much breaking, the webgl port now utilises the browser's ogg/mp3 decoder, and the browser's png decompresser. pngs/jpegs do not provide size information. fixed non-browser ogg decoding and openal logic. should finally be fixed, I guess. wasted some time on an ezhud plugin, along with ensuring certain info is available to the plugin. this is still a work in progress however. non-web builds are now able to download images from urls. download progress does not display for uri_get/uri_post builtins, nor other non-saved things. q1qvm logic now uses pr_maxedicts cvar. the gamecode api still provides no way to tell how many are permissable, so set at own risk.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4852 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/pr_csqc.c
@@ -4146,7 +4146,7 @@ typedef struct
 	unsigned int maxents;	//buffer size
 	struct
 	{
-		unsigned short n;	//don't rely on the ent->v->entnum
+		unsigned int n;	//don't rely on the ent->v->entnum
 		csqcedict_t *e;	//the csqc ent
 	} *e;
 } csqcdelta_pack_t;
